refactor(hyperswarm): organize mediator
